app/requirements.txt
streamlit>=1.29
pandas>=2.0
openpyxl>=3.1
SQLAlchemy>=2.0
psycopg2-binary>=2.9
python-dotenv>=1.0
